% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getBoardList.R
\name{getBoardList}
\alias{getBoardList}
\title{板一覧}
\usage{
getBoardList(bbsmenu.uri = "http://menu.2ch.net/bbsmenu.html")
}
\arguments{
\item{bbsmenu.uri}{メニューの URI}
}
\description{
板一覧をデータフレームとして取得します。
}
\details{
以下のデータフレームが返ります。

\code{Category}   カテゴリ名。

\code{Name}	板名。

\code{URI}	板 (ホストではない) のルート URI。
}

