% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSubjectList.R
\name{getSubjectList}
\alias{getSubjectList}
\title{スレタイ一覧}
\usage{
getSubjectList(server.uri, bg20 = FALSE)
}
\arguments{
\item{server.uri}{鯖 URI}

\item{bg20}{鯖を利用するか}
}
\description{
スレッド一覧をデータフレームとして取得します。
}
\details{
以下のデータフレームが返ります。

\code{DAT}   DAT ファイルの URI。

\code{Subject}   スレッドのタイトル。

\code{Count}	スレッドのレス数。
}

