% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getThread.R
\name{getThread}
\alias{getThread}
\title{スレ}
\usage{
getThread(dat.uri)
}
\arguments{
\item{dat.uri}{DAT ファイルの URI}
}
\description{
スレッドをデータフレームとして取得します。
時刻は 1970 年 1 月 1 日午前 9 時 0 分 0 秒からの経過秒数として表されます。
}
\details{
以下のデータフレームが返ります。

\code{Subject}   スレッドのタイトル。

\code{Number}	レス番号。

\code{Trip}	トリップ。

\code{Email}	メールアドレス。

\code{ID}	ID。

\code{BE}	BE ID。

\code{DateTime}	投稿日時。 1970-01-01 09:00:00 (UTC) からの経過秒数。

\code{Message}	本文。
}

