% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printThread.R
\name{printThread}
\alias{printThread}
\title{レス出力}
\usage{
printThread(thread, index = 1:nrow(thread), by = 1,
  format = function(post) {     email <- sprintf("[\%s]", post$Email)    
  datetime <- strftime(as.POSIXct(post$DateTime, origin =
  "1970-01-01 09:00:00"), "\%Y/\%m/\%d(\%a) \%H:\%M:\%S")     header <-
  paste(post$Number, post$Name, email, datetime)     message <- paste("--",
  header, "--", post$Message, sep = "\\n")     return(message) })
}
\arguments{
\item{thread}{スレッドデータフレーム}

\item{index}{出力するレス番号}

\item{by}{1 回に出力するレス数}

\item{format}{レスの出力フォーマット}
}
\description{
スレッドの内容をコンソールに出力します。
}

