\name{R2ch-package}
\alias{R2ch-package}
\alias{R2ch}
\docType{package}
\title{
R2ch package
}
\description{
A reader of the 2ch bulletin board.
}
\details{
\tabular{ll}{
Package: \tab R2ch\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-23\cr
License: \tab The MIT License\cr
LazyLoad: \tab yes\cr
}
}
\author{
RecycleBin

Maintainer: kos59125
}
\references{
}
\keyword{2ch}
\keyword{bulletin board}
\seealso{
}
\examples{
}
